backend/requirements_v2.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9

# NLP and ML
spacy>=3.7.0
transformers>=4.35.0
torch>=2.1.0
numpy>=1.24.0
pandas>=2.1.0
nltk>=3.8.0
scikit-learn
xgboost
lightgbm>=1.3.0
redis>=5.0.0
youtube-transcript-api>=0.6.0
keybert>=0.8.0
prophet>=1.1.5

# API and HTTP
httpx>=0.25.0
requests>=2.31.0

# Social Media APIs
tweepy>=4.14.0
praw>=7.7.0
pytrends>=4.9.0
google-search-results>=2.4.2

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
